[TE] fix(efa): make AV slot removal refcounted and inflight-aware
With the SRD shared-endpoint model every peer is an fi_addr_t index in the
per-NIC AV, and libfabric hands back the SAME index for the same peer
address while fi_av_remove() frees that index for reuse.  Two consequences
were unhandled:

1. Several EfaEndPoint handles can legitimately hold one fi_addr_t at
   once.  With a volatile peer key (ip:port:timestamp@nic) the same peer is
   re-admitted under a new key while the old handle is still alive, and
   each handshake inserts the same address.  The first handle torn down
   called fi_av_remove() unconditionally, invalidating the slot its
   siblings were still posting on -- observed as SIGSEGV inside libfabric
   from fi_write() with an in-range but freed dest_addr (insert / post /
   remove / post on one slot from three threads within 60us).

2. fi_av_remove() on a slot with posted-but-uncompleted operations makes
   the provider drop those completions, so the slices stay PENDING and the
   caller's status loop spins on WAITING forever (hang reproduced in 5 of 5
   runs with a saturated peer_map_).

Fix: per-slot AvSlotState{refcount, inflight, remove_pending} keyed by
fi_addr_t.  Both counters have to live on the slot, not the endpoint -- an
endpoint-local counter only sees ONE holder's operations, so an idle
endpoint retiring the slot still eats a sibling's completions.  The slot is
removed only when the last holder releases it AND inflight is zero;
otherwise removal is deferred and retired by the CQ poller once the slot
quiesces, so no completion is lost and no slot is leaked.  inflight is
incremented before the post and rolled back on FI_EAGAIN / hard failure.
EfaOpContext carries the slot pointer so the poller can settle it.

Also:
  * two call sites (peer_map_ eviction on failed reconnect, and warmup
    handshake failure) used the unconditional deleteEndpoint(); both now
    use deleteEndpointIfIdle() so a busy peer is left for the poller.
  * stale-peer CQ errors (prov_errno 4/8/9/14) are classified and the
    handle retired instead of failing the transfer permanently; the noisy
    per-completion logs are rate-limited with LOG_EVERY_N.
  * MC_MAX_EP_PER_CTX ceiling raised from UINT16_MAX to INT32_MAX and
    parsed with atoll -- the AV is bounded by the provider, not by 16 bits.

Validated on two real-EFA p5.48xlarge hosts (H100, 16 EFA NICs), CPU and
CUDA builds:
  * unit tests: 8 suites / 72 cases green (CPU), 5 suites green (CUDA)
  * healthy 16-NIC bench, 5 runs: CPU-to-CPU write 206-213 GB/s, read
    203 GB/s, 0 failed batches; GPU-to-GPU (gpu_id=-1, 8 GPUs) write
    292 GB/s, read 251 GB/s -- all within noise of baseline
  * pressure repro (MC_MAX_EP_PER_CTX=1, forcing constant slot reuse):
    baseline 5/5 HANG, patched 5/5 PASS; bidirectional both sides HANG ->
    both sides PASS; EPC=2 and EPC=4 HANG -> PASS

Note this fixes slot lifetime, not AV capacity: a peer key that grows
without bound still exhausts the AV (fi_av_insert ENOMEM), which needs a
key-strategy change on the application side.

+++ mooncake-transfer-engine/src/config.cpp
@@ -97,9 +97,15 @@ void loadGlobalConfig(GlobalConfig& config) {
 
     const char* max_ep_per_ctx_env = std::getenv("MC_MAX_EP_PER_CTX");
     if (max_ep_per_ctx_env) {
-        size_t val = atoi(max_ep_per_ctx_env);
-        if (val > 0 && val <= UINT16_MAX)
-            config.max_ep_per_ctx = val;
+        // Ceiling is INT32_MAX, not UINT16_MAX: the field is an int whose
+        // default is already 65536, so a UINT16_MAX bound made the default
+        // value impossible to set explicitly.  The real RDMA constraint is
+        // enforced later against device_attr.max_qp; on EFA (SRD) this only
+        // sizes the AV table and the peer_map_ eviction cap, neither of which
+        // is 16-bit.
+        long long val = atoll(max_ep_per_ctx_env);
+        if (val > 0 && val <= INT32_MAX)
+            config.max_ep_per_ctx = static_cast<int>(val);
         else
             LOG(WARNING)
                 << "Ignore value from environment variable MC_MAX_EP_PER_CTX";
